Moved mocito extensions to shared module
