+++ library/build.gradle
@@ -39,7 +39,7 @@ dependencies {
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
     api project(":library-core")
-    api 'com.squareup.okhttp3:okhttp:3.9.1'
+    api 'com.squareup.okhttp3:okhttp:3.10.0'
     api 'com.google.code.gson:gson:2.8.2'
 
     testImplementation 'junit:junit:4.12'
@@ -48,10 +48,11 @@ dependencies {
     testImplementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     testImplementation "org.jetbrains.kotlin:kotlin-test-junit:$kotlin_version"
     testImplementation "com.google.truth:truth:0.39"
+    testImplementation project(":test-common")
 
     androidTestImplementation 'com.android.support.test:runner:1.0.1'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
-    androidTestImplementation "com.google.truth:truth:0.39"
+    androidTestImplementation "com.google.truth:truth:0.40"
     androidTestImplementation 'org.mockito:mockito-android:2.10.0'
 
 }
